# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HIMSARU ERP API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Sales report with date range filter
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
  /reports/purchases:
    get:
      operationId: getPurchasesReport
      tags: [reports]
      summary: Purchases report with date range filter
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Purchases report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchasesReport"
...
  /reports/profit:
    get:
      operationId: getProfitReport
      tags: [reports]
      summary: Profit report with date range filter
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Profit report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitReport"
...
components:
  schemas:
...
    SalesReportRow:
      type: object
      properties:
        id:
          type: integer
        saleDate:
          type: string
        invoiceNo:
          type: ["string", "null"]
        customerName:
          type: string
        gstType:
          type: string
        taxableAmount:
          type: number
        discount:
          type: number
        cgstAmount:
          type: number
        sgstAmount:
          type: number
        igstAmount:
          type: number
        totalTax:
          type: number
        totalAmount:
          type: number
      required: [id, saleDate, customerName, gstType, taxableAmount, discount, cgstAmount, sgstAmount, igstAmount, totalTax, totalAmount]
    SalesReportSummary:
      type: object
      properties:
        count:
          type: integer
        totalTaxable:
          type: number
        totalDiscount:
          type: number
        totalTax:
          type: number
        totalAmount:
          type: number
      required: [count, totalTaxable, totalDiscount, totalTax, totalAmount]
    SalesReport:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/SalesReportRow"
        summary:
          $ref: "#/components/schemas/SalesReportSummary"
      required: [rows, summary]
    PurchasesReportRow:
      type: object
      properties:
        id:
          type: integer
        purchaseDate:
          type: string
        invoiceNo:
          type: ["string", "null"]
        supplierName:
          type: string
        totalAmount:
          type: number
      required: [id, purchaseDate, supplierName, totalAmount]
    PurchasesReportSummary:
      type: object
      properties:
        count:
          type: integer
        totalAmount:
          type: number
      required: [count, totalAmount]
    PurchasesReport:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/PurchasesReportRow"
        summary:
          $ref: "#/components/schemas/PurchasesReportSummary"
      required: [rows, summary]
...
    ProfitReportRow:
      type: object
      properties:
        id:
          type: integer
        saleDate:
          type: string
        invoiceNo:
          type: ["string", "null"]
        customerName:
          type: string
        revenue:
          type: number
        cogs:
          type: number
        grossProfit:
          type: number
        margin:
          type: number
        totalTax:
          type: number
        totalAmount:
          type: number
      required: [id, saleDate, customerName, revenue, cogs, grossProfit, margin, totalTax, totalAmount]
    ProfitReportSummary:
      type: object
      properties:
        count:
          type: integer
        totalRevenue:
          type: number
        totalCogs:
          type: number
        totalGrossProfit:
          type: number
        overallMargin:
          type: number
      required: [count, totalRevenue, totalCogs, totalGrossProfit, overallMargin]
    ProfitReport:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ProfitReportRow"
        summary:
          $ref: "#/components/schemas/ProfitReportSummary"
      required: [rows, summary]
...